Revert "gh-128384: Use a context variable for warnings.catch_warnings (gh-130010)"
This reverts commit d687900f98114bb5910daad9553ae381d7daf94b.
