+++ Doc/library/decimal.rst
@@ -1884,20 +1884,13 @@ the current thread.
 
 If :func:`setcontext` has not been called before :func:`getcontext`, then
 :func:`getcontext` will automatically create a new context for use in the
-current thread.  New context objects have default values set from the
-:data:`decimal.DefaultContext` object.
+current thread.
 
-The :data:`sys.flags.thread_inherit_context` flag affects the context for
-new threads.  If the flag is false, new threads will start with an empty
-context.  In this case, :func:`getcontext` will create a new context object
-when called and use the default values from *DefaultContext*.  If the flag
-is true, new threads will start with a copy of context from the caller of
-:meth:`threading.Thread.start`.
-
-To control the defaults so that each thread will use the same values throughout
-the application, directly modify the *DefaultContext* object. This should be
-done *before* any threads are started so that there won't be a race condition
-between threads calling :func:`getcontext`. For example::
+The new context is copied from a prototype context called *DefaultContext*. To
+control the defaults so that each thread will use the same values throughout the
+application, directly modify the *DefaultContext* object. This should be done
+*before* any threads are started so that there won't be a race condition between
+threads calling :func:`getcontext`. For example::
 
    # Set applicationwide defaults for all threads about to be launched
    DefaultContext.prec = 12
