+++ Include/cpython/initconfig.h
@@ -180,8 +180,6 @@ typedef struct PyConfig {
     int use_frozen_modules;
     int safe_path;
     int int_max_str_digits;
-    int thread_inherit_context;
-    int context_aware_warnings;
 #ifdef __APPLE__
     int use_system_logger;
 #endif
