+++ Include/internal/pycore_interp_structs.h
@@ -582,7 +582,6 @@ struct _warnings_runtime_state {
     PyObject *default_action; /* String */
     _PyRecursiveMutex lock;
     long filters_version;
-    PyObject *context;
 };
 
 struct _Py_mem_interp_free_queue {
