+++ Lib/test/support/warnings_helper.py
@@ -160,12 +160,11 @@ def _filterwarnings(filters, quiet=False):
     registry = frame.f_globals.get('__warningregistry__')
     if registry:
         registry.clear()
-    # Because test_warnings swap the module, we need to look up in the
-    # sys.modules dictionary.
-    wmod = sys.modules['warnings']
-    with wmod.catch_warnings(record=True) as w:
-        # Set filter "always" to record all warnings.
-        wmod.simplefilter("always")
+    with warnings.catch_warnings(record=True) as w:
+        # Set filter "always" to record all warnings.  Because
+        # test_warnings swap the module, we need to look up in
+        # the sys.modules dictionary.
+        sys.modules['warnings'].simplefilter("always")
         yield WarningsRecorder(w)
     # Filter the recorded warnings
     reraise = list(w)
