+++ Lib/test/test_capi/test_config.py
@@ -54,8 +54,6 @@ class CAPITests(unittest.TestCase):
             ("filesystem_errors", str, None),
             ("hash_seed", int, None),
             ("home", str | None, None),
-            ("thread_inherit_context", int, None),
-            ("context_aware_warnings", int, None),
             ("import_time", bool, None),
             ("inspect", bool, None),
             ("install_signal_handlers", bool, None),
@@ -100,7 +98,7 @@ class CAPITests(unittest.TestCase):
         ]
         if support.Py_DEBUG:
             options.append(("run_presite", str | None, None))
-        if support.Py_GIL_DISABLED:
+        if sysconfig.get_config_var('Py_GIL_DISABLED'):
             options.append(("enable_gil", int, None))
             options.append(("tlbc_enabled", int, None))
         if support.MS_WINDOWS:
@@ -172,7 +170,7 @@ class CAPITests(unittest.TestCase):
             ("warn_default_encoding", "warn_default_encoding", False),
             ("safe_path", "safe_path", False),
             ("int_max_str_digits", "int_max_str_digits", False),
-            # "gil", "thread_inherit_context" and "context_aware_warnings" are tested below
+            # "gil" is tested below
         ):
             with self.subTest(flag=flag, name=name, negate=negate):
                 value = config_get(name)
@@ -184,16 +182,10 @@ class CAPITests(unittest.TestCase):
                          config_get('use_hash_seed') == 0
                          or config_get('hash_seed') != 0)
 
-        if support.Py_GIL_DISABLED:
+        if sysconfig.get_config_var('Py_GIL_DISABLED'):
             value = config_get('enable_gil')
             expected = (value if value != -1 else None)
             self.assertEqual(sys.flags.gil, expected)
-
-        expected_inherit_context = 1 if support.Py_GIL_DISABLED else 0
-        self.assertEqual(sys.flags.thread_inherit_context, expected_inherit_context)
-
-        expected_safe_warnings = 1 if support.Py_GIL_DISABLED else 0
-        self.assertEqual(sys.flags.context_aware_warnings, expected_safe_warnings)
 
     def test_config_get_non_existent(self):
         # Test PyConfig_Get() on non-existent option name
