+++ Lib/test/test_context.py
@@ -1,4 +1,3 @@
-import sys
 import collections.abc
 import concurrent.futures
 import contextvars
@@ -392,60 +391,6 @@ class ContextTest(unittest.TestCase):
         finally:
             tp.shutdown()
         self.assertEqual(results, list(range(10)))
-
-    @isolated_context
-    @threading_helper.requires_working_threading()
-    def test_context_thread_inherit(self):
-        import threading
-
-        cvar = contextvars.ContextVar('cvar')
-
-        def run_context_none():
-            if sys.flags.thread_inherit_context:
-                expected = 1
-            else:
-                expected = None
-            self.assertEqual(cvar.get(None), expected)
-
-        # By default, context is inherited based on the
-        # sys.flags.thread_inherit_context option.
-        cvar.set(1)
-        thread = threading.Thread(target=run_context_none)
-        thread.start()
-        thread.join()
-
-        # Passing 'None' explicitly should have same behaviour as not
-        # passing parameter.
-        thread = threading.Thread(target=run_context_none, context=None)
-        thread.start()
-        thread.join()
-
-        # An explicit Context value can also be passed
-        custom_ctx = contextvars.Context()
-        custom_var = None
-
-        def setup_context():
-            nonlocal custom_var
-            custom_var = contextvars.ContextVar('custom')
-            custom_var.set(2)
-
-        custom_ctx.run(setup_context)
-
-        def run_custom():
-            self.assertEqual(custom_var.get(), 2)
-
-        thread = threading.Thread(target=run_custom, context=custom_ctx)
-        thread.start()
-        thread.join()
-
-        # You can also pass a new Context() object to start with an empty context
-        def run_empty():
-            with self.assertRaises(LookupError):
-                cvar.get()
-
-        thread = threading.Thread(target=run_empty, context=contextvars.Context())
-        thread.start()
-        thread.join()
 
     def test_token_contextmanager_with_default(self):
         ctx = contextvars.Context()
