+++ Lib/test/test_decimal.py
@@ -45,7 +45,6 @@ from test.support import warnings_helper
 import random
 import inspect
 import threading
-import contextvars
 
 
 if sys.platform == 'darwin':
@@ -1732,13 +1731,8 @@ class ThreadingTest:
         self.finish1 = threading.Event()
         self.finish2 = threading.Event()
 
-        # This test wants to start threads with an empty context, no matter
-        # the setting of sys.flags.thread_inherit_context.  We pass the
-        # 'context' argument explicitly with an empty context instance.
-        th1 = threading.Thread(target=thfunc1, args=(self,),
-                               context=contextvars.Context())
-        th2 = threading.Thread(target=thfunc2, args=(self,),
-                               context=contextvars.Context())
+        th1 = threading.Thread(target=thfunc1, args=(self,))
+        th2 = threading.Thread(target=thfunc2, args=(self,))
 
         th1.start()
         th2.start()
