+++ Lib/test/test_warnings/__init__.py
@@ -24,13 +24,10 @@ import warnings as original_warnings
 from warnings import deprecated
 
 
-py_warnings = import_helper.import_fresh_module('_py_warnings')
-py_warnings._set_module(py_warnings)
-
-c_warnings = import_helper.import_fresh_module(
-    "warnings", fresh=["_warnings", "_py_warnings"]
-)
-c_warnings._set_module(c_warnings)
+py_warnings = import_helper.import_fresh_module('warnings',
+                                                blocked=['_warnings'])
+c_warnings = import_helper.import_fresh_module('warnings',
+                                               fresh=['_warnings'])
 
 @contextmanager
 def warnings_state(module):
@@ -46,21 +43,15 @@ def warnings_state(module):
     except NameError:
         pass
     original_warnings = warning_tests.warnings
-    if module._use_context:
-        saved_context, context = module._new_context()
-    else:
-        original_filters = module.filters
-        module.filters = original_filters[:]
+    original_filters = module.filters
     try:
+        module.filters = original_filters[:]
         module.simplefilter("once")
         warning_tests.warnings = module
         yield
     finally:
         warning_tests.warnings = original_warnings
-        if module._use_context:
-            module._set_context(saved_context)
-        else:
-            module.filters = original_filters
+        module.filters = original_filters
 
 
 class TestWarning(Warning):
@@ -120,14 +111,14 @@ class FilterTests(BaseTest):
     """Testing the filtering functionality."""
 
     def test_error(self):
-        with self.module.catch_warnings() as w:
+        with original_warnings.catch_warnings(module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("error", category=UserWarning)
             self.assertRaises(UserWarning, self.module.warn,
                                 "FilterTests.test_error")
 
     def test_error_after_default(self):
-        with self.module.catch_warnings() as w:
+        with original_warnings.catch_warnings(module=self.module) as w:
             self.module.resetwarnings()
             message = "FilterTests.test_ignore_after_default"
             def f():
@@ -145,7 +136,8 @@ class FilterTests(BaseTest):
             self.assertRaises(UserWarning, f)
 
     def test_ignore(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("ignore", category=UserWarning)
             self.module.warn("FilterTests.test_ignore", UserWarning)
@@ -153,7 +145,8 @@ class FilterTests(BaseTest):
             self.assertEqual(list(__warningregistry__), ['version'])
 
     def test_ignore_after_default(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             message = "FilterTests.test_ignore_after_default"
             def f():
@@ -166,7 +159,8 @@ class FilterTests(BaseTest):
 
     def test_always_and_all(self):
         for mode in {"always", "all"}:
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.resetwarnings()
                 self.module.filterwarnings(mode, category=UserWarning)
                 message = "FilterTests.test_always_and_all"
@@ -181,7 +175,8 @@ class FilterTests(BaseTest):
 
     def test_always_and_all_after_default(self):
         for mode in {"always", "all"}:
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.resetwarnings()
                 message = "FilterTests.test_always_and_all_after_ignore"
                 def f():
@@ -200,7 +195,8 @@ class FilterTests(BaseTest):
                 self.assertEqual(w[-1].message.args[0], message)
 
     def test_default(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("default", category=UserWarning)
             message = UserWarning("FilterTests.test_default")
@@ -215,7 +211,8 @@ class FilterTests(BaseTest):
                     raise ValueError("loop variant unhandled")
 
     def test_module(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("module", category=UserWarning)
             message = UserWarning("FilterTests.test_module")
@@ -226,7 +223,8 @@ class FilterTests(BaseTest):
             self.assertEqual(len(w), 0)
 
     def test_once(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("once", category=UserWarning)
             message = UserWarning("FilterTests.test_once")
@@ -242,7 +240,8 @@ class FilterTests(BaseTest):
             self.assertEqual(len(w), 0)
 
     def test_module_globals(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.simplefilter("always", UserWarning)
 
             # bpo-33509: module_globals=None must not crash
@@ -262,14 +261,15 @@ class FilterTests(BaseTest):
             self.assertEqual(len(w), 2)
 
     def test_inheritance(self):
-        with self.module.catch_warnings() as w:
+        with original_warnings.catch_warnings(module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("error", category=Warning)
             self.assertRaises(UserWarning, self.module.warn,
                                 "FilterTests.test_inheritance", UserWarning)
 
     def test_ordering(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("ignore", category=UserWarning)
             self.module.filterwarnings("error", category=UserWarning,
@@ -284,7 +284,8 @@ class FilterTests(BaseTest):
     def test_filterwarnings(self):
         # Test filterwarnings().
         # Implicitly also tests resetwarnings().
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.filterwarnings("error", "", Warning, "", 0)
             self.assertRaises(UserWarning, self.module.warn, 'convert to error')
 
@@ -308,7 +309,8 @@ class FilterTests(BaseTest):
             self.assertIs(w[-1].category, UserWarning)
 
     def test_message_matching(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.simplefilter("ignore", UserWarning)
             self.module.filterwarnings("error", "match", UserWarning)
             self.assertRaises(UserWarning, self.module.warn, "match")
@@ -324,52 +326,54 @@ class FilterTests(BaseTest):
                 L[:] = []
 
         L = [("default",X(),UserWarning,X(),0) for i in range(2)]
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.filters = L
             self.module.warn_explicit(UserWarning("b"), None, "f.py", 42)
             self.assertEqual(str(w[-1].message), "b")
 
     def test_filterwarnings_duplicate_filters(self):
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.resetwarnings()
             self.module.filterwarnings("error", category=UserWarning)
-            self.assertEqual(len(self.module._get_filters()), 1)
+            self.assertEqual(len(self.module.filters), 1)
             self.module.filterwarnings("ignore", category=UserWarning)
             self.module.filterwarnings("error", category=UserWarning)
             self.assertEqual(
-                len(self.module._get_filters()), 2,
+                len(self.module.filters), 2,
                 "filterwarnings inserted duplicate filter"
             )
             self.assertEqual(
-                self.module._get_filters()[0][0], "error",
+                self.module.filters[0][0], "error",
                 "filterwarnings did not promote filter to "
                 "the beginning of list"
             )
 
     def test_simplefilter_duplicate_filters(self):
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.resetwarnings()
             self.module.simplefilter("error", category=UserWarning)
-            self.assertEqual(len(self.module._get_filters()), 1)
+            self.assertEqual(len(self.module.filters), 1)
             self.module.simplefilter("ignore", category=UserWarning)
             self.module.simplefilter("error", category=UserWarning)
             self.assertEqual(
-                len(self.module._get_filters()), 2,
+                len(self.module.filters), 2,
                 "simplefilter inserted duplicate filter"
             )
             self.assertEqual(
-                self.module._get_filters()[0][0], "error",
+                self.module.filters[0][0], "error",
                 "simplefilter did not promote filter to the beginning of list"
             )
 
     def test_append_duplicate(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(module=self.module,
+                record=True) as w:
             self.module.resetwarnings()
             self.module.simplefilter("ignore")
             self.module.simplefilter("error", append=True)
             self.module.simplefilter("ignore", append=True)
             self.module.warn("test_append_duplicate", category=UserWarning)
-            self.assertEqual(len(self.module._get_filters()), 2,
+            self.assertEqual(len(self.module.filters), 2,
                 "simplefilter inserted duplicate filter"
             )
             self.assertEqual(len(w), 0,
@@ -399,17 +403,19 @@ class FilterTests(BaseTest):
             self.module.simplefilter('ignore', lineno=-1)
 
     def test_catchwarnings_with_simplefilter_ignore(self):
-        with self.module.catch_warnings(module=self.module):
+        with original_warnings.catch_warnings(module=self.module):
             self.module.resetwarnings()
             self.module.simplefilter("error")
-            with self.module.catch_warnings(action="ignore"):
+            with self.module.catch_warnings(
+                module=self.module, action="ignore"
+            ):
                 self.module.warn("This will be ignored")
 
     def test_catchwarnings_with_simplefilter_error(self):
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.resetwarnings()
             with self.module.catch_warnings(
-                action="error", category=FutureWarning
+                module=self.module, action="error", category=FutureWarning
             ):
                 with support.captured_stderr() as stderr:
                     error_msg = "Other types of warnings are not errors"
@@ -431,7 +437,8 @@ class WarnTests(BaseTest):
     """Test warnings.warn() and warnings.warn_explicit()."""
 
     def test_message(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.simplefilter("once")
             for i in range(4):
                 text = 'multi %d' %i  # Different text on each call.
@@ -443,7 +450,8 @@ class WarnTests(BaseTest):
     def test_warn_nonstandard_types(self):
         # warn() should handle non-standard types without issue.
         for ob in (Warning, None, 42):
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.simplefilter("once")
                 self.module.warn(ob)
                 # Don't directly compare objects since
@@ -452,7 +460,8 @@ class WarnTests(BaseTest):
 
     def test_filename(self):
         with warnings_state(self.module):
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 warning_tests.inner("spam1")
                 self.assertEqual(os.path.basename(w[-1].filename),
                                     "stacklevel.py")
@@ -464,7 +473,8 @@ class WarnTests(BaseTest):
         # Test stacklevel argument
         # make sure all messages are different, so the warning won't be skipped
         with warnings_state(self.module):
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 warning_tests.inner("spam3", stacklevel=1)
                 self.assertEqual(os.path.basename(w[-1].filename),
                                     "stacklevel.py")
@@ -490,7 +500,8 @@ class WarnTests(BaseTest):
         # Issue #24305: With stacklevel=2, module-level warnings should work.
         import_helper.unload('test.test_warnings.data.import_warning')
         with warnings_state(self.module):
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.simplefilter('always')
                 import test.test_warnings.data.import_warning  # noqa: F401
                 self.assertEqual(len(w), 1)
@@ -498,7 +509,8 @@ class WarnTests(BaseTest):
 
     def test_skip_file_prefixes(self):
         with warnings_state(self.module):
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.simplefilter('always')
 
                 # Warning never attributed to the data/ package.
@@ -525,7 +537,9 @@ class WarnTests(BaseTest):
         # see: gh-126209
         with warnings_state(self.module):
             skipped = warning_tests.__file__
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(
+                record=True, module=self.module,
+            ) as w:
                 warning_tests.outer("msg", skip_file_prefixes=(skipped,))
 
             self.assertEqual(len(w), 1)
@@ -546,13 +560,14 @@ class WarnTests(BaseTest):
         codeobj = compile(("import warnings\n"
                            "warnings.warn('hello', UserWarning)"),
                           filename, "exec")
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True) as w:
             self.module.simplefilter("always", category=UserWarning)
             exec(codeobj)
         self.assertEqual(w[0].filename, filename)
 
     def test_warn_explicit_non_ascii_filename(self):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(record=True,
+                module=self.module) as w:
             self.module.resetwarnings()
             self.module.filterwarnings("always", category=UserWarning)
             filenames = ["nonascii\xe9\u20ac"]
@@ -622,7 +637,7 @@ class WarnTests(BaseTest):
         self.assertIn('category must be a Warning subclass, not ',
                       str(cm.exception))
 
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.resetwarnings()
             self.module.filterwarnings('default')
             with self.assertWarns(MyWarningClass) as cm:
@@ -638,7 +653,7 @@ class WarnTests(BaseTest):
             self.assertIsInstance(cm.warning, Warning)
 
     def check_module_globals(self, module_globals):
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(module=self.module, record=True) as w:
             self.module.filterwarnings('default')
             self.module.warn_explicit(
                 'eggs', UserWarning, 'bar', 1,
@@ -651,7 +666,7 @@ class WarnTests(BaseTest):
         if self.module is py_warnings:
             self.check_module_globals(module_globals)
             return
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(module=self.module, record=True) as w:
             self.module.filterwarnings('always')
             with self.assertRaisesRegex(errtype, re.escape(errmsg)):
                 self.module.warn_explicit(
@@ -663,7 +678,7 @@ class WarnTests(BaseTest):
         if self.module is py_warnings:
             self.check_module_globals(module_globals)
             return
-        with self.module.catch_warnings(record=True) as w:
+        with original_warnings.catch_warnings(module=self.module, record=True) as w:
             self.module.filterwarnings('always')
             self.module.warn_explicit(
                 'eggs', UserWarning, 'bar', 1,
@@ -752,7 +767,7 @@ class WCmdLineTests(BaseTest):
     def test_improper_input(self):
         # Uses the private _setoption() function to test the parsing
         # of command-line warning arguments
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.assertRaises(self.module._OptionError,
                               self.module._setoption, '1:2:3:4:5:6')
             self.assertRaises(self.module._OptionError,
@@ -771,7 +786,7 @@ class WCmdLineTests(BaseTest):
             self.assertRaises(UserWarning, self.module.warn, 'convert to error')
 
     def test_import_from_module(self):
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module._setoption('ignore::Warning')
             with self.assertRaises(self.module._OptionError):
                 self.module._setoption('ignore::TestWarning')
@@ -814,7 +829,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
 
     def test_filter(self):
         # Everything should function even if 'filters' is not in warnings.
-        with self.module.catch_warnings() as w:
+        with original_warnings.catch_warnings(module=self.module) as w:
             self.module.filterwarnings("error", "", Warning, "", 0)
             self.assertRaises(UserWarning, self.module.warn,
                                 'convert to error')
@@ -829,7 +844,8 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         try:
             original_registry = self.module.onceregistry
             __warningregistry__ = {}
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.resetwarnings()
                 self.module.filterwarnings("once", category=UserWarning)
                 self.module.warn_explicit(message, UserWarning, "file", 42)
@@ -856,7 +872,8 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         message = UserWarning("defaultaction test")
         original = self.module.defaultaction
         try:
-            with self.module.catch_warnings(record=True) as w:
+            with original_warnings.catch_warnings(record=True,
+                    module=self.module) as w:
                 self.module.resetwarnings()
                 registry = {}
                 self.module.warn_explicit(message, UserWarning, "<test>", 42,
@@ -889,12 +906,8 @@ class _WarningsTests(BaseTest, unittest.TestCase):
 
     def test_showwarning_missing(self):
         # Test that showwarning() missing is okay.
-        if self.module._use_context:
-            # If _use_context is true, the warnings module does not
-            # override/restore showwarning()
-            return
         text = 'del showwarning test'
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.filterwarnings("always", category=UserWarning)
             del self.module.showwarning
             with support.captured_output('stderr') as stream:
@@ -905,7 +918,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
     def test_showwarnmsg_missing(self):
         # Test that _showwarnmsg() missing is okay.
         text = 'del _showwarnmsg test'
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.filterwarnings("always", category=UserWarning)
 
             show = self.module._showwarnmsg
@@ -919,7 +932,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         self.assertIn(text, result)
 
     def test_showwarning_not_callable(self):
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.filterwarnings("always", category=UserWarning)
             self.module.showwarning = print
             with support.captured_output('stdout'):
@@ -930,7 +943,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
     def test_show_warning_output(self):
         # With showwarning() missing, make sure that output is okay.
         text = 'test show_warning'
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.filterwarnings("always", category=UserWarning)
             del self.module.showwarning
             with support.captured_output('stderr') as stream:
@@ -955,11 +968,12 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         globals_dict = globals()
         oldfile = globals_dict['__file__']
         try:
-            catch = self.module.catch_warnings(record=True)
+            catch = original_warnings.catch_warnings(record=True,
+                                                     module=self.module)
             with catch as w:
                 self.module.filterwarnings("always", category=UserWarning)
                 globals_dict['__file__'] = None
-                self.module.warn('test', UserWarning)
+                original_warnings.warn('test', UserWarning)
                 self.assertTrue(len(w))
         finally:
             globals_dict['__file__'] = oldfile
@@ -996,7 +1010,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
 
 
         wmod = self.module
-        with wmod.catch_warnings():
+        with original_warnings.catch_warnings(module=wmod):
             wmod.filterwarnings('default', category=UserWarning)
 
             linecache.clearcache()
@@ -1023,7 +1037,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         # warn_explicit() shouldn't raise a SystemError in case
         # warnings.onceregistry isn't a dictionary.
         wmod = self.module
-        with wmod.catch_warnings():
+        with original_warnings.catch_warnings(module=wmod):
             wmod.filterwarnings('once')
             with support.swap_attr(wmod, 'onceregistry', None):
                 with self.assertRaises(TypeError):
@@ -1034,12 +1048,12 @@ class _WarningsTests(BaseTest, unittest.TestCase):
         # warn_explicit() shouldn't cause an assertion failure in case of a
         # bad warnings.filters or warnings.defaultaction.
         wmod = self.module
-        with wmod.catch_warnings():
-            wmod._get_filters()[:] = [(None, None, Warning, None, 0)]
+        with original_warnings.catch_warnings(module=wmod):
+            wmod.filters = [(None, None, Warning, None, 0)]
             with self.assertRaises(TypeError):
                 wmod.warn_explicit('foo', Warning, 'bar', 1)
 
-            wmod._get_filters()[:] = []
+            wmod.filters = []
             with support.swap_attr(wmod, 'defaultaction', None), \
                  self.assertRaises(TypeError):
                 wmod.warn_explicit('foo', Warning, 'bar', 1)
@@ -1048,7 +1062,7 @@ class _WarningsTests(BaseTest, unittest.TestCase):
     def test_issue31566(self):
         # warn() shouldn't cause an assertion failure in case of a bad
         # __name__ global.
-        with self.module.catch_warnings():
+        with original_warnings.catch_warnings(module=self.module):
             self.module.filterwarnings('error', category=UserWarning)
             with support.swap_item(globals(), '__name__', b'foo'), \
                  support.swap_item(globals(), '__file__', None):
@@ -1176,18 +1190,16 @@ class CatchWarningTests(BaseTest):
     """Test catch_warnings()."""
 
     def test_catch_warnings_restore(self):
-        if self.module._use_context:
-            return  # test disabled if using context vars
         wmod = self.module
         orig_filters = wmod.filters
         orig_showwarning = wmod.showwarning
         # Ensure both showwarning and filters are restored when recording
-        with wmod.catch_warnings(record=True):
+        with wmod.catch_warnings(module=wmod, record=True):
             wmod.filters = wmod.showwarning = object()
         self.assertIs(wmod.filters, orig_filters)
         self.assertIs(wmod.showwarning, orig_showwarning)
         # Same test, but with recording disabled
-        with wmod.catch_warnings(record=False):
+        with wmod.catch_warnings(module=wmod, record=False):
             wmod.filters = wmod.showwarning = object()
         self.assertIs(wmod.filters, orig_filters)
         self.assertIs(wmod.showwarning, orig_showwarning)
@@ -1195,7 +1207,7 @@ class CatchWarningTests(BaseTest):
     def test_catch_warnings_recording(self):
         wmod = self.module
         # Ensure warnings are recorded when requested
-        with wmod.catch_warnings(record=True) as w:
+        with wmod.catch_warnings(module=wmod, record=True) as w:
             self.assertEqual(w, [])
             self.assertIs(type(w), list)
             wmod.simplefilter("always")
@@ -1209,48 +1221,44 @@ class CatchWarningTests(BaseTest):
             self.assertEqual(w, [])
         # Ensure warnings are not recorded when not requested
         orig_showwarning = wmod.showwarning
-        with wmod.catch_warnings(record=False) as w:
+        with wmod.catch_warnings(module=wmod, record=False) as w:
             self.assertIsNone(w)
             self.assertIs(wmod.showwarning, orig_showwarning)
 
     def test_catch_warnings_reentry_guard(self):
         wmod = self.module
         # Ensure catch_warnings is protected against incorrect usage
-        x = wmod.catch_warnings(record=True)
+        x = wmod.catch_warnings(module=wmod, record=True)
         self.assertRaises(RuntimeError, x.__exit__)
         with x:
             self.assertRaises(RuntimeError, x.__enter__)
         # Same test, but with recording disabled
-        x = wmod.catch_warnings(record=False)
+        x = wmod.catch_warnings(module=wmod, record=False)
         self.assertRaises(RuntimeError, x.__exit__)
         with x:
             self.assertRaises(RuntimeError, x.__enter__)
 
     def test_catch_warnings_defaults(self):
         wmod = self.module
-        orig_filters = wmod._get_filters()
+        orig_filters = wmod.filters
         orig_showwarning = wmod.showwarning
         # Ensure default behaviour is not to record warnings
-        with wmod.catch_warnings() as w:
+        with wmod.catch_warnings(module=wmod) as w:
             self.assertIsNone(w)
             self.assertIs(wmod.showwarning, orig_showwarning)
-            self.assertIsNot(wmod._get_filters(), orig_filters)
-        self.assertIs(wmod._get_filters(), orig_filters)
+            self.assertIsNot(wmod.filters, orig_filters)
+        self.assertIs(wmod.filters, orig_filters)
         if wmod is sys.modules['warnings']:
             # Ensure the default module is this one
             with wmod.catch_warnings() as w:
                 self.assertIsNone(w)
                 self.assertIs(wmod.showwarning, orig_showwarning)
-                self.assertIsNot(wmod._get_filters(), orig_filters)
-            self.assertIs(wmod._get_filters(), orig_filters)
+                self.assertIsNot(wmod.filters, orig_filters)
+            self.assertIs(wmod.filters, orig_filters)
 
     def test_record_override_showwarning_before(self):
         # Issue #28835: If warnings.showwarning() was overridden, make sure
         # that catch_warnings(record=True) overrides it again.
-        if self.module._use_context:
-            # If _use_context is true, the warnings module does not restore
-            # showwarning()
-            return
         text = "This is a warning"
         wmod = self.module
         my_log = []
@@ -1261,7 +1269,7 @@ class CatchWarningTests(BaseTest):
 
         # Override warnings.showwarning() before calling catch_warnings()
         with support.swap_attr(wmod, 'showwarning', my_logger):
-            with wmod.catch_warnings(record=True) as log:
+            with wmod.catch_warnings(module=wmod, record=True) as log:
                 self.assertIsNot(wmod.showwarning, my_logger)
 
                 wmod.simplefilter("always")
@@ -1276,10 +1284,6 @@ class CatchWarningTests(BaseTest):
     def test_record_override_showwarning_inside(self):
         # Issue #28835: It is possible to override warnings.showwarning()
         # in the catch_warnings(record=True) context manager.
-        if self.module._use_context:
-            # If _use_context is true, the warnings module does not restore
-            # showwarning()
-            return
         text = "This is a warning"
         wmod = self.module
         my_log = []
@@ -1288,7 +1292,7 @@ class CatchWarningTests(BaseTest):
             nonlocal my_log
             my_log.append(message)
 
-        with wmod.catch_warnings(record=True) as log:
+        with wmod.catch_warnings(module=wmod, record=True) as log:
             wmod.simplefilter("always")
             wmod.showwarning = my_logger
             wmod.warn(text)
@@ -1402,7 +1406,7 @@ class EnvironmentVariableTests(BaseTest):
             code = "import sys; sys.modules.pop('warnings', None); sys.modules['_warnings'] = None; "
         else:
             code = ""
-        code += "import warnings; [print(f) for f in warnings._get_filters()]"
+        code += "import warnings; [print(f) for f in warnings.filters]"
 
         rc, stdout, stderr = assert_python_ok("-c", code, __isolated=True)
         stdout_lines = [line.strip() for line in stdout.splitlines()]
@@ -1526,169 +1530,6 @@ a=A()
         code = "import warnings; f = open(%a)" % __file__
         rc, out, err = assert_python_ok("-Wd", "-c", code)
         self.assertTrue(err.startswith(expected), ascii(err))
-
-
-class AsyncTests(BaseTest):
-    """Verifies that the catch_warnings() context manager behaves
-    as expected when used inside async co-routines.  This requires
-    that the context_aware_warnings flag is enabled, so that
-    the context manager uses a context variable.
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.module.resetwarnings()
-
-    @unittest.skipIf(not sys.flags.context_aware_warnings,
-                     "requires context aware warnings")
-    def test_async_context(self):
-        import asyncio
-
-        # Events to force the execution interleaving we want.
-        step_a1 = asyncio.Event()
-        step_a2 = asyncio.Event()
-        step_b1 = asyncio.Event()
-        step_b2 = asyncio.Event()
-
-        async def run_a():
-            with self.module.catch_warnings(record=True) as w:
-                await step_a1.wait()
-                # The warning emitted here should be caught be the enclosing
-                # context manager.
-                self.module.warn('run_a warning', UserWarning)
-                step_b1.set()
-                await step_a2.wait()
-                self.assertEqual(len(w), 1)
-                self.assertEqual(w[0].message.args[0], 'run_a warning')
-                step_b2.set()
-
-        async def run_b():
-            with self.module.catch_warnings(record=True) as w:
-                step_a1.set()
-                await step_b1.wait()
-                # The warning emitted here should be caught be the enclosing
-                # context manager.
-                self.module.warn('run_b warning', UserWarning)
-                step_a2.set()
-                await step_b2.wait()
-                self.assertEqual(len(w), 1)
-                self.assertEqual(w[0].message.args[0], 'run_b warning')
-
-        async def run_tasks():
-            await asyncio.gather(run_a(), run_b())
-
-        asyncio.run(run_tasks())
-
-    @unittest.skipIf(not sys.flags.context_aware_warnings,
-                     "requires context aware warnings")
-    def test_async_task_inherit(self):
-        """Check that a new asyncio task inherits warnings context from the
-        coroutine that spawns it.
-        """
-        import asyncio
-
-        step1 = asyncio.Event()
-        step2 = asyncio.Event()
-
-        async def run_child1():
-            await step1.wait()
-            # This should be recorded by the run_parent() catch_warnings
-            # context.
-            self.module.warn('child warning', UserWarning)
-            step2.set()
-
-        async def run_child2():
-            # This establishes a new catch_warnings() context.  The
-            # run_child1() task should still be using the context from
-            # run_parent() if context-aware warnings are enabled.
-            with self.module.catch_warnings(record=True) as w:
-                step1.set()
-                await step2.wait()
-
-        async def run_parent():
-            with self.module.catch_warnings(record=True) as w:
-                await asyncio.gather(run_child1(), run_child2())
-                self.assertEqual(len(w), 1)
-                self.assertEqual(w[0].message.args[0], 'child warning')
-
-        asyncio.run(run_parent())
-
-
-class CAsyncTests(AsyncTests, unittest.TestCase):
-    module = c_warnings
-
-
-class PyAsyncTests(AsyncTests, unittest.TestCase):
-    module = py_warnings
-
-
-class ThreadTests(BaseTest):
-    """Verifies that the catch_warnings() context manager behaves as
-    expected when used within threads.  This requires that both the
-    context_aware_warnings flag and thread_inherit_context flags are enabled.
-    """
-
-    ENABLE_THREAD_TESTS = (sys.flags.context_aware_warnings and
-                           sys.flags.thread_inherit_context)
-
-    def setUp(self):
-        super().setUp()
-        self.module.resetwarnings()
-
-    @unittest.skipIf(not ENABLE_THREAD_TESTS,
-                     "requires thread-safe warnings flags")
-    def test_threaded_context(self):
-        import threading
-
-        barrier = threading.Barrier(2)
-
-        def run_a():
-            with self.module.catch_warnings(record=True) as w:
-                barrier.wait()
-                # The warning emitted here should be caught be the enclosing
-                # context manager.
-                self.module.warn('run_a warning', UserWarning)
-                barrier.wait()
-                self.assertEqual(len(w), 1)
-                self.assertEqual(w[0].message.args[0], 'run_a warning')
-            # Should be caught be the catch_warnings() context manager of run_threads()
-            self.module.warn('main warning', UserWarning)
-
-        def run_b():
-            with self.module.catch_warnings(record=True) as w:
-                barrier.wait()
-                # The warning emitted here should be caught be the enclosing
-                # context manager.
-                barrier.wait()
-                self.module.warn('run_b warning', UserWarning)
-                self.assertEqual(len(w), 1)
-                self.assertEqual(w[0].message.args[0], 'run_b warning')
-            # Should be caught be the catch_warnings() context manager of run_threads()
-            self.module.warn('main warning', UserWarning)
-
-        def run_threads():
-            threads = [
-                threading.Thread(target=run_a),
-                threading.Thread(target=run_b),
-                ]
-            with self.module.catch_warnings(record=True) as w:
-                for thread in threads:
-                    thread.start()
-                for thread in threads:
-                    thread.join()
-                self.assertEqual(len(w), 2)
-                self.assertEqual(w[0].message.args[0], 'main warning')
-                self.assertEqual(w[1].message.args[0], 'main warning')
-
-        run_threads()
-
-
-class CThreadTests(ThreadTests, unittest.TestCase):
-    module = c_warnings
-
-
-class PyThreadTests(ThreadTests, unittest.TestCase):
-    module = py_warnings
 
 
 class DeprecatedTests(PyPublicAPITests):
