--- Misc/NEWS.d/next/Core_and_Builtins/2025-01-06-10-55-41.gh-issue-128555.tAK_AY.rst
@@ -1,16 +1,0 @@
-Add the :data:`sys.flags.thread_inherit_context` flag.
-
-* This flag is set to true by default on the free-threaded build
-  and false otherwise.  If the flag is true, starting a new thread using
-  :class:`threading.Thread` will, by default, use a copy of the
-  :class:`contextvars.Context` from the caller of
-  :meth:`threading.Thread.start` rather than using an empty context.
-
-* Add the :option:`-X thread_inherit_context <-X>` command-line option and
-  :envvar:`PYTHON_THREAD_INHERIT_CONTEXT` environment variable, which set the
-  :data:`~sys.flags.thread_inherit_context` flag.
-
-* Add the ``context`` keyword parameter to :class:`~threading.Thread`.  It can
-  be used to explicitly pass a context value to be used by a new thread.
-
-* Make the ``_contextvars`` module built-in.
