+++ Python/stdlib_module_names.h
@@ -64,7 +64,6 @@ static const char* _Py_stdlib_module_names[] = {
 "_posixshmem",
 "_posixsubprocess",
 "_py_abc",
-"_py_warnings",
 "_pydatetime",
 "_pydecimal",
 "_pyio",
