+++ configure
@@ -805,6 +805,8 @@ MODULE__HEAPQ_FALSE
 MODULE__HEAPQ_TRUE
 MODULE__CSV_FALSE
 MODULE__CSV_TRUE
+MODULE__CONTEXTVARS_FALSE
+MODULE__CONTEXTVARS_TRUE
 MODULE__BISECT_FALSE
 MODULE__BISECT_TRUE
 MODULE__ASYNCIO_FALSE
@@ -30769,6 +30771,28 @@ then :
 fi
 
 
+        if test "$py_cv_module__contextvars" != "n/a"
+then :
+  py_cv_module__contextvars=yes
+fi
+   if test "$py_cv_module__contextvars" = yes; then
+  MODULE__CONTEXTVARS_TRUE=
+  MODULE__CONTEXTVARS_FALSE='#'
+else
+  MODULE__CONTEXTVARS_TRUE='#'
+  MODULE__CONTEXTVARS_FALSE=
+fi
+
+  as_fn_append MODULE_BLOCK "MODULE__CONTEXTVARS_STATE=$py_cv_module__contextvars$as_nl"
+  if test "x$py_cv_module__contextvars" = xyes
+then :
+
+
+
+
+fi
+
+
         if test "$py_cv_module__csv" != "n/a"
 then :
   py_cv_module__csv=yes
@@ -33677,6 +33701,10 @@ Usually this means the macro was only invoked conditionally." "$LINENO" 5
 fi
 if test -z "${MODULE__BISECT_TRUE}" && test -z "${MODULE__BISECT_FALSE}"; then
   as_fn_error $? "conditional \"MODULE__BISECT\" was never defined.
+Usually this means the macro was only invoked conditionally." "$LINENO" 5
+fi
+if test -z "${MODULE__CONTEXTVARS_TRUE}" && test -z "${MODULE__CONTEXTVARS_FALSE}"; then
+  as_fn_error $? "conditional \"MODULE__CONTEXTVARS\" was never defined.
 Usually this means the macro was only invoked conditionally." "$LINENO" 5
 fi
 if test -z "${MODULE__CSV_TRUE}" && test -z "${MODULE__CSV_FALSE}"; then
